[Merged] reservation new and reservation index page
