+++ src/assets/styles/app.less
@@ -880,3 +880,16 @@ body #app {
 .list .item-content {
   padding-right: 0.85rem;
 }
+
+.reservation-form .list {
+  margin: 0;
+
+  ul::before,
+  ul::after {
+    content: none;
+  }
+}
+
+.min-h-20 {
+  min-height: 5rem;
+}
