+++ src/common/routes/index.ts
@@ -7,6 +7,7 @@ import SignUpPage from '@pages/users/registrations/new';
 import IntroPage from '@pages/intro';
 import UserEditPage from '@pages/users/registrations/edit';
 import MyPage from '@pages/mypage';
+import ReservationNewPage from '@pages/reservations/new';
 
 /**
  * @resourceRoutes
@@ -26,10 +27,8 @@ const resourceRoutes: ResourceRoute[] = [
     resource: 'users',
   },
   { resource: 'groups' },
-  { resource: 'reservations' },
   { resource: 'application_forms' },
   { resource: 'contacts' },
-
 ];
 
 /**
@@ -44,6 +43,7 @@ const customRoutes = [
   { path: '/users/sign_up', component: SignUpPage },
   { path: '/users/edit', component: UserEditPage },
   { path: '/mypage', component: MyPage },
+  { path: '/reservations', component: ReservationNewPage },
 ];
 
 /**
